Claude/fsr thermistor adc console mhjciw

+++ tfg_test_firmware/components/sensor_manager/include/imu_driver.h
@@ -3,18 +3,23 @@
 #include "driver/i2c_master.h"
 #include "esp_err.h"
 
+/*
+ * Valores RAW directos de los registros del LSM9DS1 (sin convertir a
+ * unidades fisicas). La conversion (sensibilidad g/LSB, dps/LSB,
+ * gauss/LSB) se hace en la app movil -- ver sensor_calibration.dart.
+ */
 typedef struct {
-  float ax;
-  float ay;
-  float az;
+  int16_t ax;
+  int16_t ay;
+  int16_t az;
 
-  float gx;
-  float gy;
-  float gz;
+  int16_t gx;
+  int16_t gy;
+  int16_t gz;
 
-  float mx;
-  float my;
-  float mz;
+  int16_t mx;
+  int16_t my;
+  int16_t mz;
 } imu_data_t;
 
 esp_err_t imu_driver_init(i2c_master_bus_handle_t bus);
